requirements.txt
Django>=1.7
python-dateutil==2.4.2
